main/chembo/requirements.txt
absl-py==0.7.1
astor==0.7.1
certifi==2019.3.9
cycler==0.10.0
Cython==0.29.7
decorator==4.4.0
dragonfly-opt==0.1.2
future==0.17.1
gast==0.2.2
grpcio==1.20.1
h5py==2.9.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
Markdown==3.1
matplotlib==3.0.3
mkl-fft==1.0.12
mkl-random==1.0.2
mock==2.0.0
mpi4py==1.3.1
networkx==2.3
numpy==1.16.3
olefile==0.46
pandas==0.22.0
pbr==5.2.0
Pillow==6.2.0
pkgconfig==1.5.1
POT==0.5.1
protobuf==3.7.1
pyparsing==2.4.0
python-dateutil==2.8.0
python-igraph==0.7.1.post6
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
seaborn==0.9.0
six==1.12.0
sklearn==0.0
tensorboard==1.13.1
tensorflow==1.15.2
tensorflow-estimator==1.13.0
termcolor==1.1.0
tqdm==4.31.1
Werkzeug==0.15.3
